update for evaluation on github

+++ exercises/14_class_destruct/main.cpp
@@ -6,7 +6,7 @@
 /// @brief 任意缓存容量的斐波那契类型。
 /// @details 可以在构造时传入缓存容量，因此需要动态分配缓存空间。
 class DynFibonacci {
-    size_t *cache;
+    unsigned long long *cache;
     int cached;
 
 public:
